+++ scripts/carride/nodes/powerup_slow_motion.gd
@@ -1,6 +1,0 @@
-extends PowerupClass
-
-
-func _ready() -> void:
-	type = Powerups.SLOWMOTION
-	time = 10