+++ scripts/subsystems/audio_stream_player.gd
@@ -6,14 +6,15 @@ enum BtnType {
 }
 
 var musicfiles = [
-	preload("res://audio/music/bloo-stricken_commision.ogg"),
+	preload("res://audio/music/stricken-commision.ogg"),
 	preload("res://audio/music/phonky.ogg"),
 	preload("res://audio/music/digitist.ogg"),
 	preload("res://audio/music/tech-junkie.ogg"),
 	preload("res://audio/music/dead-ahead.ogg"),
 	preload("res://audio/music/hard-trouble.ogg"),
 	preload("res://audio/music/sahara.ogg"),
-	preload("res://audio/music/bloo-good_luck.ogg"),
+	preload("res://audio/music/good-luck.ogg"),
+	preload("res://audio/music/cyber-punk.ogg"),
 ]
 
 var soundeffects = [
@@ -23,7 +24,6 @@ var soundeffects = [
 	preload("res://audio/sfx/button_on.wav"),
 ]
 
-var stricken_remastered = preload("res://audio/music/stricken_commision_remastered.ogg")
 var random: int = 0
 var spooky_easter_egg := false
 
@@ -33,12 +33,13 @@ func _enter_tree() -> void:
 	if current_date['month'] == Time.MONTH_OCTOBER and current_date['day'] == 31:
 		spooky_easter_egg = true
 		stream = load("res://audio/music/spooky-scary-skeletons.ogg")
+		musicfiles.append(stream)
 		play()
 
 
 func connect_buttons(root):
 	for child in root.get_children():
-		if not child.is_in_group("NoAudio"):
+		if not child.is_in_group("NoButtonPressSFX"):
 			if child is BaseButton and child.toggle_mode == true:
 				child.toggled.connect(button_sound.bind(BtnType.SWITCH))
 			elif child is BaseButton:
@@ -47,18 +48,9 @@ func connect_buttons(root):
 
 
 func change_track():
-	if not spooky_easter_egg:
-		if random == 0:
-			var rnd_remaster = randi_range(0,10)
-			if rnd_remaster == 0:
-				var pos = get_playback_position()
-				stream = stricken_remastered
-				play()
-				seek(pos)
-				return
-		random = randi_range(0,len(musicfiles)-1)
-		stream = musicfiles[random]
-		play()
+	random = randi_range(0,len(musicfiles)-1)
+	stream = musicfiles[random]
+	play()
 
 
 func button_sound(on: bool, type: BtnType) -> void:
@@ -73,5 +65,4 @@ func button_sound(on: bool, type: BtnType) -> void:
 				soundeffects[1]
 		BtnType.SWITCH:
 			player.stream = soundeffects[3] if on else soundeffects[2]
-	print(player.stream)
 	player.play()